Remoção e imprementação dos códigos finais da API

+++ main/java/lojaAPI/Cliente.java
@@ -1,23 +1,26 @@
 package lojaAPI;
 
+
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+
 
 @Entity
 public class Cliente {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id_cliente;
+	private int id_cliente;
 	private String nome;
 	private String email;
 	private String telefone;
-	public Long getId_cliente() {
+	
+	public int getId_cliente() {
 		return id_cliente;
 	}
-	public void setId_cliente(Long id_cliente) {
+	public void setId_cliente(int id_cliente) {
 		this.id_cliente = id_cliente;
 	}
 	public String getNome() {
